OPENEUROPA-2391: Reorganizing classes of plugins.
