+++ src/Plugin/Block/EcFooterBlock.php
@@ -5,58 +5,19 @@ declare(strict_types = 1);
 namespace Drupal\oe_corporate_blocks\Plugin\Block;
 
 use Drupal\Component\Utility\NestedArray;
-use Drupal\Core\Block\BlockBase;
 use Drupal\Core\Cache\CacheableMetadata;
-use Drupal\Core\Config\ConfigFactoryInterface;
 use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Provides the corporate footer block.
  *
  * @Block(
  *   id = "oe_footer",
- *   admin_label = @Translation("Footer block"),
+ *   admin_label = @Translation("EC Footer block"),
  *   category = @Translation("Corporate blocks"),
  * )
  */
-class FooterBlock extends BlockBase implements ContainerFactoryPluginInterface {
-
-  /**
-   * The config factory.
-   *
-   * @var \Drupal\Core\Config\ConfigFactoryInterface
-   */
-  protected $configFactory;
-
-  /**
-   * Construct the footer block object.
-   *
-   * @param array $configuration
-   *   A configuration array containing information about the plugin instance.
-   * @param string $plugin_id
-   *   The plugin_id for the plugin instance.
-   * @param mixed $plugin_definition
-   *   The plugin implementation definition.
-   * @param \Drupal\Core\Config\ConfigFactoryInterface $config_factory
-   *   The config factory.
-   */
-  public function __construct(array $configuration, $plugin_id, $plugin_definition, ConfigFactoryInterface $config_factory) {
-    parent::__construct($configuration, $plugin_id, $plugin_definition);
-    $this->configFactory = $config_factory;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
-    return new static(
-      $configuration,
-      $plugin_id,
-      $plugin_definition,
-      $container->get('config.factory')
-    );
-  }
+class EcFooterBlock extends FooterBlockBase implements ContainerFactoryPluginInterface {
 
   /**
    * {@inheritdoc}
